combine the `include` tests into 1 file
it seems a little overkill to have each compile test in a separate file, so we'll move them into 1 file. This, IMO is easier to comprehend, because now we can easily make sure they're all testing the same thing.

This test file then covers the entirety of the `CompilesIncludes.php` file.
